Split HDLightEditor file
