+++ com.unity.render-pipelines.high-definition/Editor/Lighting/HDLightEditor.ContextMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9e6395c99e5b6b4ab6769e2270d8f88
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
