+++ com.unity.render-pipelines.high-definition/Editor/Lighting/HDLightEditor.Drawers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 652fe95f0fd8dec458c380d86458890d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
