+++ com.unity.render-pipelines.high-definition/Editor/Lighting/HDLightEditor.Handles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9e030f1ae3cd6c343a945303f593ebe8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
