+++ com.unity.render-pipelines.high-definition/Editor/Lighting/HDLightEditor.Serialized.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0546b569891a2424d86539c0c4fa5404
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
